Implementation of KNNRegressor as sklearn-model
