--- docs/drevalpy.models.KNNRegressor.rst
@@ -1,9 +1,0 @@
-KNNRegressor
-========
-
-KNNRegressor Model
----------------
-.. automodule:: drevalpy.models.KNNRegressor.drug_gnn
-   :members:
-   :undoc-members:
-   :show-inheritance:
