+++ docs/drevalpy.models.rst
@@ -31,4 +31,3 @@ Implemented models
    drevalpy.models.SimpleNeuralNetwork
    drevalpy.models.SuperFELTR
    drevalpy.models.baselines
-   drevalpy.models.KNNRegressor
