--- drevalpy/models/KNNRegressor/__init__.py
@@ -1,5 +1,0 @@
-"""KNNRegressor model."""
-
-from .knn_regressor import KNNRegressor
-
-__all__ = ["KNNRegressor"]
