+++ drevalpy/models/baselines/hyperparameters.yaml
@@ -202,3 +202,19 @@ AdaBoostDecisionTree:
   min_samples_leaf:
     - 1
     - 3
+KNNRegressor:
+  n_neighbors:
+    - 5
+    - 10
+    - 3
+  weights:
+    - distance
+  variance:
+    - 0.75
+    - 0.8
+    - 0.85
+    - 0.9
+    - 0.95
+    - 0.7
+    - 0.6
+    - 0.5
