+++ tests/test_drp_model.py
@@ -43,7 +43,6 @@ def test_factory() -> None:
     assert "MOLIR" in MODEL_FACTORY
     assert "SuperFELTR" in MODEL_FACTORY
     assert "DIPK" in MODEL_FACTORY
-    assert "KNNRegressor" in MODEL_FACTORY
 
 
 def test_load_cl_ids_from_csv() -> None:
